requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
scikit-image==0.22.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.25.2
Pillow==10.1.0
PyYAML==6.0.1
pytesseract==0.3.10
pyzbar==0.1.9
ImageHash==4.3.1
requests==2.31.0
Jinja2==3.1.2
reportlab==4.0.7
